requirements.txt
predmarket
requests
python-dotenv
httpx
pydantic
google-generativeai
